Add Goomba Activation
This was such btch to imp.
Its not like the logic itself was hard.
What was hard dwas imp. within an existing
code base, which is something I only recently
started to really deal with. Premature
optimizaiton was also a btch, which is actually
why I created Notes.txt to keep track of
optimzation that may be good ideas for future
me to deal with.